E2E: Move Sync and CL logic to their own files
