+++ tests/e2e/helpers_e2e_test.go
@@ -13,63 +13,10 @@ import (
 	"github.com/osmosis-labs/osmosis/v19/tests/e2e/configurer/chain"
 	"github.com/osmosis-labs/osmosis/v19/tests/e2e/initialization"
 	"github.com/osmosis-labs/osmosis/v19/tests/e2e/util"
-	"github.com/osmosis-labs/osmosis/v19/x/concentrated-liquidity/model"
-	"github.com/osmosis-labs/osmosis/v19/x/concentrated-liquidity/types"
 	gammtypes "github.com/osmosis-labs/osmosis/v19/x/gamm/types"
 )
 
 var defaultFeePerTx = osmomath.NewInt(1000)
-
-// calculateSpreadRewardGrowthGlobal calculates spread reward growth global per unit of virtual liquidity based on swap parameters:
-// amountIn - amount being swapped
-// spreadFactor - pool's spread factor
-// poolLiquidity - current pool liquidity
-func calculateSpreadRewardGrowthGlobal(spreadRewardChargeTotal, poolLiquidity osmomath.Dec) osmomath.Dec {
-	// First we get total spread reward charge for the swap (ΔY * spreadFactor)
-
-	// Calculating spread reward growth global (dividing by pool liquidity to find spread reward growth per unit of virtual liquidity)
-	spreadRewardGrowthGlobal := spreadRewardChargeTotal.QuoTruncate(poolLiquidity)
-	return spreadRewardGrowthGlobal
-}
-
-// calculateSpreadRewardGrowthInside calculates spread reward growth inside range per unit of virtual liquidity
-// spreadRewardGrowthGlobal - global spread reward growth per unit of virtual liquidity
-// spreadRewardGrowthBelow - spread reward growth below lower tick
-// spreadRewardGrowthAbove - spread reward growth above upper tick
-// Formula: spreadRewardGrowthGlobal - spreadRewardGrowthBelowLowerTick - spreadRewardGrowthAboveUpperTick
-func calculateSpreadRewardGrowthInside(spreadRewardGrowthGlobal, spreadRewardGrowthBelow, spreadRewardGrowthAbove osmomath.Dec) osmomath.Dec {
-	return spreadRewardGrowthGlobal.Sub(spreadRewardGrowthBelow).Sub(spreadRewardGrowthAbove)
-}
-
-// Assert balances that are not affected by swap:
-// * same amount of `stake` in balancesBefore and balancesAfter
-// * amount of `e2e-default-feetoken` dropped by 1000 (default amount for fee per tx)
-// * depending on `assertUosmoBalanceIsConstant` and `assertUionBalanceIsConstant` parameters, check that those balances have also not been changed
-func (s *IntegrationTestSuite) assertBalancesInvariants(balancesBefore, balancesAfter sdk.Coins, assertUosmoBalanceIsConstant, assertUionBalanceIsConstant bool) {
-	s.Require().True(balancesAfter.AmountOf("stake").Equal(balancesBefore.AmountOf("stake")))
-	s.Require().True(balancesAfter.AmountOf("e2e-default-feetoken").Equal(balancesBefore.AmountOf("e2e-default-feetoken").Sub(defaultFeePerTx)))
-	if assertUionBalanceIsConstant {
-		s.Require().True(balancesAfter.AmountOf("uion").Equal(balancesBefore.AmountOf("uion")))
-	}
-	if assertUosmoBalanceIsConstant {
-		s.Require().True(balancesAfter.AmountOf("uosmo").Equal(balancesBefore.AmountOf("uosmo")))
-	}
-}
-
-func (s *IntegrationTestSuite) assertClSwap(clpoolStart types.ConcentratedPoolExtension, clPoolAfter types.ConcentratedPoolExtension, expectedSqrtPriceDelta osmomath.BigDec) {
-	// Update pool and track liquidity and sqrt price
-	liquidityBeforeSwap := clpoolStart.GetLiquidity()
-	sqrtPriceBeforeSwap := clpoolStart.GetCurrentSqrtPrice()
-	liquidityAfterSwap := clPoolAfter.GetLiquidity()
-	sqrtPriceAfterSwap := clPoolAfter.GetCurrentSqrtPrice()
-
-	// Assert swaps don't change pool's liquidity amount
-	s.Require().Equal(liquidityAfterSwap.String(), liquidityBeforeSwap.String())
-
-	// Assert current sqrt price
-	expectedSqrtPrice := sqrtPriceBeforeSwap.Add(expectedSqrtPriceDelta)
-	s.Require().Equal(expectedSqrtPrice.String(), sqrtPriceAfterSwap.String())
-}
 
 // Get balances for address
 func (s *IntegrationTestSuite) addrBalance(node *chain.NodeConfig, address string) sdk.Coins {
@@ -128,13 +75,6 @@ func calculateUncollectedSpreadRewards(positionLiquidity, spreadRewardGrowthBelo
 	return spreadRewardsUncollected
 }
 
-// Get current (updated) pool
-func (s *IntegrationTestSuite) updatedConcentratedPool(node *chain.NodeConfig, poolId uint64) types.ConcentratedPoolExtension {
-	concentratedPool, err := node.QueryConcentratedPool(poolId)
-	s.Require().NoError(err)
-	return concentratedPool
-}
-
 func (s *IntegrationTestSuite) updatedCFMMPool(node *chain.NodeConfig, poolId uint64) gammtypes.CFMMPoolI {
 	cfmmPool, err := node.QueryCFMMPool(poolId)
 	s.Require().NoError(err)
@@ -146,13 +86,6 @@ func formatCLIInt(i int) string {
 		return fmt.Sprintf("[%d]", i)
 	}
 	return strconv.Itoa(i)
-}
-
-// Assert returned positions:
-func (s *IntegrationTestSuite) validateCLPosition(position model.Position, poolId uint64, lowerTick, upperTick int64) {
-	s.Require().Equal(position.PoolId, poolId)
-	s.Require().Equal(position.LowerTick, lowerTick)
-	s.Require().Equal(position.UpperTick, upperTick)
 }
 
 func (s *IntegrationTestSuite) CallCheckBalance(node *chain.NodeConfig, addr, denom string, amount int64) {
