+++ tests/e2e/sync_test.go
@@ -1,0 +1,110 @@
+package e2e
+
+import (
+	"fmt"
+	"os"
+	"path/filepath"
+	"time"
+
+	coretypes "github.com/tendermint/tendermint/rpc/core/types"
+
+	"github.com/osmosis-labs/osmosis/v19/tests/e2e/initialization"
+)
+
+// TODO: Consider moving this to its own package and having separate test instantiation for it.
+func (s *IntegrationTestSuite) StateSync() {
+	if s.skipStateSync {
+		s.T().Skip()
+	}
+
+	// This test benefits from the use of chainA's default node, since it has
+	// the shortest snapshot interval.
+	chainA := s.configurer.GetChainConfig(0)
+	chainANode, err := chainA.GetDefaultNode()
+	s.Require().NoError(err)
+
+	persistentPeers := chainA.GetPersistentPeers()
+
+	stateSyncHostPort := fmt.Sprintf("%s:26657", chainANode.Name)
+	stateSyncRPCServers := []string{stateSyncHostPort, stateSyncHostPort}
+
+	// get trust height and trust hash.
+	trustHeight, err := chainANode.QueryCurrentHeight()
+	s.Require().NoError(err)
+
+	trustHash, err := chainANode.QueryHashFromBlock(trustHeight)
+	s.Require().NoError(err)
+
+	stateSynchingNodeConfig := &initialization.NodeConfig{
+		Name:               "state-sync",
+		Pruning:            "default",
+		PruningKeepRecent:  "0",
+		PruningInterval:    "0",
+		SnapshotInterval:   1500,
+		SnapshotKeepRecent: 2,
+	}
+
+	tempDir, err := os.MkdirTemp("", "osmosis-e2e-statesync-")
+	s.Require().NoError(err)
+
+	// configure genesis and config files for the state-synchin node.
+	nodeInit, err := initialization.InitSingleNode(
+		chainA.Id,
+		tempDir,
+		filepath.Join(chainANode.ConfigDir, "config", "genesis.json"),
+		stateSynchingNodeConfig,
+		time.Duration(chainA.VotingPeriod),
+		// time.Duration(chainA.ExpeditedVotingPeriod),
+		trustHeight,
+		trustHash,
+		stateSyncRPCServers,
+		persistentPeers,
+	)
+	s.Require().NoError(err)
+
+	// Call tempNode method here to not add the node to the list of nodes.
+	// This messes with the nodes running in parallel if we add it to the regular list.
+	stateSynchingNode := chainA.CreateNodeTemp(nodeInit)
+
+	// ensure that the running node has snapshots at a height > trustHeight.
+	hasSnapshotsAvailable := func(syncInfo coretypes.SyncInfo) bool {
+		snapshotHeight := chainANode.SnapshotInterval
+		if uint64(syncInfo.LatestBlockHeight) < snapshotHeight {
+			s.T().Logf("snapshot height is not reached yet, current (%d), need (%d)", syncInfo.LatestBlockHeight, snapshotHeight)
+			return false
+		}
+
+		snapshots, err := chainANode.QueryListSnapshots()
+		s.Require().NoError(err)
+
+		for _, snapshot := range snapshots {
+			if snapshot.Height > uint64(trustHeight) {
+				s.T().Log("found state sync snapshot after trust height")
+				return true
+			}
+		}
+		s.T().Log("state sync snashot after trust height is not found")
+		return false
+	}
+	chainANode.WaitUntil(hasSnapshotsAvailable)
+
+	// start the state synchin node.
+	err = stateSynchingNode.Run()
+	s.Require().NoError(err)
+
+	// ensure that the state synching node cathes up to the running node.
+	s.Require().Eventually(func() bool {
+		stateSyncNodeHeight, err := stateSynchingNode.QueryCurrentHeight()
+		s.Require().NoError(err)
+		runningNodeHeight, err := chainANode.QueryCurrentHeight()
+		s.Require().NoError(err)
+		return stateSyncNodeHeight == runningNodeHeight
+	},
+		1*time.Minute,
+		10*time.Millisecond,
+	)
+
+	// stop the state synching node.
+	err = chainA.RemoveTempNode(stateSynchingNode.Name)
+	s.Require().NoError(err)
+}
